Some final changes to the Front End
