+++ apps/web/src/services/jobService.ts
@@ -5,7 +5,7 @@ const API_URL = process.env.REACT_APP_BACKEND_URL;
 export const fetchJobs = async () => {
   try {
     const response = await axios.get(`${API_URL}/jobs`);
-    console.log(response.data);
+    // console.log(response.data);
     return response.data;
   } catch (error) {
     console.error("Failed to fetch jobs:", error);
